refactor: Add `User` prefix to component names
"Settings" is very generic. This will give us problems later, with importing components in the router file.
