+++ src/router/index.ts
@@ -2,15 +2,15 @@
 import 'vue-router';
 import { createRouter, createWebHistory } from 'vue-router';
 import Home from '../views/Home.vue';
-import Login from '../views/user/Login.vue';
-import Register from '../views/user/Register.vue';
-import Settings from '../views/user/Settings.vue';
+import UserLogin from '../views/user/UserLogin.vue';
+import UserRegister from '../views/user/UserRegister.vue';
+import UserSettings from '../views/user/UserSettings.vue';
 
 const routes = [
     { path: '/', component: Home },
-    { path: '/login', component: Login },
-    { path: '/register', component: Register },
-    { path: '/update', component: Settings },
+    { path: '/login', component: UserLogin },
+    { path: '/register', component: UserRegister },
+    { path: '/update', component: UserSettings },
 ];
 
 export const router = createRouter({
